♻ refactor: 상세,공유 페이지 스타일 컴포넌트 컨벤션 적용
